changes to home page (#6)
* changes to home page

Added a shared folder with certain things that will be helpful shared across the site. Added a navbar component and a few styling things. More work to be done but thats what I got done for now!

* Worked on Alex's comments and some more

Worked on more things with the home page. Alignments, different images, and more.