Features/homepage (#9)
* changes to home page

Added a shared folder with certain things that will be helpful shared across the site. Added a navbar component and a few styling things. More work to be done but thats what I got done for now!

* Worked on Alex's comments and some more

Worked on more things with the home page. Alignments, different images, and more.

* Small changes to page

Made changes to color on the page as well as navbar

* Retrying to commit new nav bar changed

Changed up the nav bar some and colors. Retrying to do this commit.

* added venv

* made changes following alexs comment

Moved Navbar to _document.tsx

* Delete api/venv directory

We should keep the Virtual Environment out of git. I forgot to add it back to git ignore after trying to get it to work for everyone.

* Create .gitignore for api folder

* feat: created dashboard (#10)

* feat: created footer (#11)

* Delete api/venv directory

* moved navbar again

* feat: remove venv

* changes to home page

Added a shared folder with certain things that will be helpful shared across the site. Added a navbar component and a few styling things. More work to be done but thats what I got done for now!

* Worked on Alex's comments and some more

Worked on more things with the home page. Alignments, different images, and more.

* Small changes to page

Made changes to color on the page as well as navbar

* Retrying to commit new nav bar changed

Changed up the nav bar some and colors. Retrying to do this commit.

* made changes following alexs comment

Moved Navbar to _document.tsx

* Delete api/venv directory

We should keep the Virtual Environment out of git. I forgot to add it back to git ignore after trying to get it to work for everyone.

* moved navbar again

* Updates and changes on homepage

* Delete yarn.lock

---------

Co-authored-by: Alex Flynn <aflynn501@gmail.com>
Co-authored-by: Alex M. Flynn <106895945+AlexMFlynn@users.noreply.github.com>
Co-authored-by: wilts <wiltsie.connor@gmail.com>