Changes to the popdown of the user profile menu bar (#17)
* Changes to the popdown of the user profile menu bar

I made a good amount of changes. First, I added login and sign up when the user clicks on the avatar menu. I made the modals work for when the user clicks either of those. I also went into signup and login .tsx and made it so when the user clicks "already a user" or "Not a user" it pops up the opposing form to fill out. There is a bug with this however that I will need help with. I also added the contact us and about us to that avatar menu bar. With the links working for those as well.

* quick fixes